# Conflicts:
#	src/FormulaSolution.java

+++ src/FormulaSolution.java
@@ -21,9 +21,8 @@ public class FormulaSolution {
         //circle area calculation
         area = Math.PI * Math.pow(radius,2);
 
-        // circle volume calculation
-//        volume = (4.0/3.0) * Math.PI * Math.pow(radius,3);
-        volume = ((double)4/3) * Math.PI * Math.pow(radius,3);
+        // sphere volume calculation
+        volume = (4.0/3.0) * Math.PI * Math.pow(radius,3);
 
         //showing answers
         System.out.println("The circumference of the circle is " + circumference);
